Add confirmation for remove playlist

+++ www/js/pages/list/list.controller.js
@@ -4,9 +4,9 @@
 
     angular.module("pages.list").controller("ListController", listController);
 
-    listController.$inject = ["$scope", "PlaylistService", "$timeout", "$ionicListDelegate"];
+    listController.$inject = ["$scope", "PlaylistService", "$timeout", "$ionicListDelegate", "$ionicPopup"];
 
-    function listController($scope, playlistService, $timeout, $ionicListDelegate) {
+    function listController($scope, playlistService, $timeout, $ionicListDelegate, $ionicPopup) {
 
       $scope.remove = remove;
       $scope.activate = activate;
@@ -14,8 +14,20 @@
       $scope.$on("$ionicView.enter", activate);
 
       function remove(item) {
-        $scope.playlists = playlistService.remove(item);
-        $ionicListDelegate.closeOptionButtons();
+
+        var confirmPopup = $ionicPopup.confirm({
+          title: 'Remove playlist',
+          template: 'Are you sure you want to remove ' + item.title + '?'
+        });
+
+        confirmPopup.then(function(res) {
+          if(res) {
+            $scope.playlists = playlistService.remove(item);
+          }
+        }).finally(function() {
+          $ionicListDelegate.closeOptionButtons();
+        });
+
       }
 
       function activate() {
